requirements.txt
absl-py==2.3.1
altair==5.5.0
altgraph==0.17.4
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
cors==1.0.1
cycler==0.12.1
defusedxml==0.7.1
Django==5.1.3
django-cors-headers==4.6.0
django-redis==6.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
fpdf==1.7.2
fpdf2==2.8.3
fsspec==2025.3.2
future==1.0.0
gast==0.6.0
gevent==24.11.1
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.74.0
h5py==3.14.0
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.11.3
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.4.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mysqlclient==2.2.6
namex==0.1.0
narwhals==1.35.0
networkx==3.4.2
numpy==2.2.4
oauthlib==3.3.1
opt_einsum==3.4.0
optree==0.17.0
packaging==24.2
panda==0.3.1
pandas==2.2.3
pefile==2023.2.7
pillow==11.0.0
playwright==1.52.0
protobuf==5.29.4
pyarrow==19.0.1
pycparser==2.22
pydeck==0.9.1
pyee==13.0.0
Pygments==2.19.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.3
PyJWT==2.10.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
razorpay==1.4.2
redis==7.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
rich==14.1.0
rpds-py==0.24.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==69.0.3
six==1.17.0
smmap==5.0.2
soupsieve==2.7
sqlparse==0.5.2
streamlit==1.44.1
sympy==1.13.1
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
threadpoolctl==3.6.0
tldextract==5.1.3
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
typing_extensions==4.13.2
tzdata==2024.2
urllib3==2.3.0
watchdog==6.0.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
zope.event==5.0
zope.interface==7.2